refactor sc eth client
