+++ nil/services/synccommittee/Makefile.inc
@@ -26,6 +26,9 @@ $(root_sc)/core/rollupcontract/rollupcontract_abi_generated.go: \
 	$(root_sc)/core/rollupcontract/abi.json
 	go generate $(root_sc)/core/rollupcontract/generate.go
 
+$(root_sc)/internal/l1client/eth_client_generated_mock.go:
+	cd $(root_sc)/internal/l1client && bash ../../internal/scripts/generate_mock.sh EthClient
+
 .PHONY: sync_committee_types_stringer
 sync_committee_types_stringer: \
 	$(root_sc)/core/batches/constraints/checkresult_string.go \
@@ -58,7 +61,7 @@ sync_committee_generate_mocks: \
 	$(root_sc)/internal/api/task_request_handler_generated_mock.go \
 	$(root_sc)/core/state_reset_launcher_generated_mock.go \
 	$(root_sc)/internal/scheduler/task_scheduler_generated_mock.go \
-	$(root_sc)/core/rollupcontract/eth_client_generated_mock.go \
+	$(root_sc)/internal/l1client/eth_client_generated_mock.go \
 	$(root_sc)/internal/srv/worker_generated_mock.go \
 	$(root_sc)/internal/test_utils/op_context_generated_mock.go \
 	$(root_sc)/prover/tracer/storage_getter_setter_generated_mock.go \
